change return type in template to `u64`

+++ template.rs
@@ -1,8 +1,8 @@
-fn part1(input: &str) -> u32 {
+fn part1(input: &str) -> u64 {
     todo!();
 }
 
-fn part2(input: &str) -> u32 {
+fn part2(input: &str) -> u64 {
     todo!();
 }
